requirements.txt
langchain
langchain-openai
langgraph
gradio
pydantic
python-dotenv
pandas
tiktoken